refactor(editor): extract preview, embeds, completion, code-block, HTML converter types from EditorView.swift (#253)
Pure file moves, no logic changes.
