+++ macOS/SynapseNotes/EditorView.swift
@@ -4273,1849 +4273,3 @@ class LinkAwareTextView: NSTextView {
         return structural.contains { lower.contains($0) }
     }
 }
-
-// MARK: - HTML to Markdown Converter
-
-/// Converts HTML content to Markdown using NSAttributedString for correct parsing,
-/// then walks the attribute runs to emit Markdown syntax.
-struct HTMLToMarkdownConverter {
-
-    static func convert(_ html: String) -> String {
-        let trimmed = html.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty else { return "" }
-
-        // Wrap in a minimal HTML document so NSAttributedString's HTML renderer
-        // uses the correct charset and a neutral sans-serif stylesheet. Without
-        // the wrapper the renderer can misdetect encoding and apply a monospace /
-        // code-block stylesheet to the entire content.
-        let wrapped: String
-        if trimmed.lowercased().hasPrefix("<!doctype") || trimmed.lowercased().hasPrefix("<html") {
-            wrapped = trimmed
-        } else {
-            wrapped = """
-            <!DOCTYPE html>
-            <html><head><meta charset="UTF-8">
-            <style>body { font-family: -apple-system, sans-serif; font-size: 13px; }</style>
-            </head><body>\(trimmed)</body></html>
-            """
-        }
-
-        guard let data = wrapped.data(using: .utf8) else { return trimmed }
-
-        let opts: [NSAttributedString.DocumentReadingOptionKey: Any] = [
-            .documentType: NSAttributedString.DocumentType.html,
-            .characterEncoding: String.Encoding.utf8.rawValue,
-        ]
-
-        guard let attrStr = try? NSAttributedString(data: data, options: opts, documentAttributes: nil) else {
-            return trimmed
-        }
-
-        return markdownFromAttributedString(attrStr)
-    }
-
-    // MARK: - Attributed string → Markdown
-
-    private static func markdownFromAttributedString(_ attrStr: NSAttributedString) -> String {
-        let fullString = attrStr.string
-        var output = ""
-
-        let nsString = fullString as NSString
-        var paraStart = 0
-
-        while paraStart < nsString.length {
-            var paraEnd = 0
-            var contentsEnd = 0
-            nsString.getParagraphStart(nil, end: &paraEnd, contentsEnd: &contentsEnd,
-                                       for: NSRange(location: paraStart, length: 0))
-
-            let contentsRange = NSRange(location: paraStart, length: contentsEnd - paraStart)
-
-            // Grab first-character attributes to classify the paragraph.
-            let attrs = (paraEnd > paraStart)
-                ? attrStr.attributes(at: paraStart, effectiveRange: nil)
-                : [:]
-
-            let paraStyle = attrs[.paragraphStyle] as? NSParagraphStyle
-            let font      = attrs[.font] as? NSFont
-            let fontSize  = font?.pointSize ?? 12
-            let headingLevel = headingLevelForFontSize(fontSize)
-            let isListItem   = isListItemParagraph(paraStyle)
-            let isOrdered    = isOrderedListItem(attrStr, range: contentsRange)
-
-            // Build inline content, suppressing bold on headings (NSAttributedString
-            // makes heading text bold by default — that would double-format it).
-            let inlineContent = inlineMarkdown(attrStr, range: contentsRange,
-                                               suppressBold: headingLevel > 0)
-                .trimmingCharacters(in: .whitespacesAndNewlines)
-
-            if inlineContent.isEmpty {
-                if !output.isEmpty { output += "\n" }
-            } else if headingLevel > 0 {
-                let hashes = String(repeating: "#", count: headingLevel)
-                output += "\(hashes) \(inlineContent)\n\n"
-            } else if isListItem {
-                let marker = isOrdered ? "1." : "-"
-                let indent = indentForParagraphStyle(paraStyle)
-                output += "\(indent)\(marker) \(inlineContent)\n"
-            } else {
-                output += "\(inlineContent)\n\n"
-            }
-
-            paraStart = paraEnd
-        }
-
-        return output
-            .replacingOccurrences(of: "\n{3,}", with: "\n\n", options: .regularExpression)
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-    }
-
-    // MARK: - Inline span rendering
-
-    private static func inlineMarkdown(_ attrStr: NSAttributedString, range: NSRange,
-                                       suppressBold: Bool = false) -> String {
-        guard range.length > 0 else { return "" }
-
-        var output = ""
-
-        attrStr.enumerateAttributes(in: range, options: []) { attrs, spanRange, _ in
-            let text = (attrStr.string as NSString).substring(with: spanRange)
-
-            // Strip tabs (list marker column) and Unicode bullets NSAttributedString
-            // inserts for <ul> items (U+2022 •, U+25E6 ◦, U+25AA ▪, etc.)
-            var cleaned = text.replacingOccurrences(of: "\t", with: "")
-            cleaned = cleaned.unicodeScalars.filter { scalar in
-                // Drop Unicode list-marker bullet characters
-                ![0x2022, 0x25E6, 0x25AA, 0x25AB, 0x2023, 0x2043].contains(scalar.value)
-            }.reduce("") { $0 + String($1) }
-
-            guard !cleaned.isEmpty else { return }
-
-            let font    = attrs[.font] as? NSFont
-            let isBold  = !suppressBold && (font?.fontDescriptor.symbolicTraits.contains(.bold) ?? false)
-            let isItalic = font?.fontDescriptor.symbolicTraits.contains(.italic) ?? false
-            let isMono  = font?.fontDescriptor.symbolicTraits.contains(.monoSpace) ?? false
-            let link    = attrs[.link] as? URL
-                       ?? (attrs[.link] as? String).flatMap { URL(string: $0) }
-
-            var span = cleaned
-
-            if isMono {
-                span = "`\(span)`"
-            } else {
-                if isBold && isItalic { span = "***\(span)***" }
-                else if isBold        { span = "**\(span)**" }
-                else if isItalic      { span = "_\(span)_" }
-            }
-
-            if let url = link, !isMono {
-                span = "[\(cleaned)](\(url.absoluteString))"
-            }
-
-            output += span
-        }
-
-        return output
-    }
-
-    // MARK: - Helpers
-
-    /// Map NSAttributedString's rendered font sizes back to heading levels.
-    /// Empirical values on macOS 14/15 with default system HTML stylesheet:
-    ///   h1 → ~24pt, h2 → ~18pt, h3 → ~14pt bold, h4-h6 → 12pt bold
-    private static func headingLevelForFontSize(_ size: CGFloat) -> Int {
-        switch size {
-        case 22...: return 1
-        case 17..<22: return 2
-        case 14..<17: return 3
-        default: return 0
-        }
-    }
-
-    private static func isListItemParagraph(_ style: NSParagraphStyle?) -> Bool {
-        guard let style else { return false }
-        return style.headIndent > 0 && !style.tabStops.isEmpty
-    }
-
-    private static func isOrderedListItem(_ attrStr: NSAttributedString, range: NSRange) -> Bool {
-        guard range.length > 0 else { return false }
-        let raw = (attrStr.string as NSString).substring(with: range)
-        // Ordered items start with a tab followed by a digit and a period.
-        return raw.hasPrefix("\t") && raw.dropFirst().first?.isNumber == true
-    }
-
-    private static func indentForParagraphStyle(_ style: NSParagraphStyle?) -> String {
-        guard let style, style.headIndent > 36 else { return "" }
-        let extraLevels = Int((style.headIndent - 18) / 18)
-        return String(repeating: "    ", count: max(0, extraLevels))
-    }
-}
-
-struct InlineImageMatch {
-    let id: String
-    let range: NSRange
-    let paragraphRange: NSRange
-    let source: String
-    let caption: String
-}
-
-struct InlineEmbedMatch {
-    let id: String
-    let range: NSRange
-    let paragraphRange: NSRange
-    let noteName: String
-    let content: String?
-    let noteURL: URL?
-}
-
-// MARK: - Embedded Notes Data Model
-
-/// Information about an embedded note for the side panel
-struct EmbeddedNoteInfo: Identifiable, Equatable {
-    let id: String
-    let noteName: String
-    let content: String?
-    let noteURL: URL?
-    let isUnresolved: Bool
-
-    static func == (lhs: EmbeddedNoteInfo, rhs: EmbeddedNoteInfo) -> Bool {
-        lhs.id == rhs.id &&
-        lhs.noteName == rhs.noteName &&
-        lhs.content == rhs.content &&
-        lhs.noteURL == rhs.noteURL &&
-        lhs.isUnresolved == rhs.isUnresolved
-    }
-}
-
-// MARK: - Unified Sidebar Embed Model
-
-/// The type of content embedded in the sidebar
-enum SidebarEmbedType {
-    case note
-    case image
-}
-
-/// Unified information about any embed (note or image) for the sidebar
-struct SidebarEmbedInfo: Identifiable, Equatable {
-    let id: String
-    let type: SidebarEmbedType
-    let title: String?       // For notes (note name)
-    let caption: String?     // For images (caption text)
-    let content: String?     // For notes (note content)
-    let source: String?      // For images (URL/path string)
-    let resolvedURL: URL?    // Resolved URL for both notes and images
-    let isUnresolved: Bool
-    let range: NSRange      // Position in document for sorting
-
-    /// Creates a SidebarEmbedInfo from an InlineEmbedMatch (note embed)
-    static func fromEmbedMatch(_ match: InlineEmbedMatch) -> SidebarEmbedInfo {
-        SidebarEmbedInfo(
-            id: match.id,
-            type: .note,
-            title: match.noteName,
-            caption: nil,
-            content: match.content,
-            source: nil,
-            resolvedURL: match.noteURL,
-            isUnresolved: match.noteURL == nil,
-            range: match.range
-        )
-    }
-
-    /// Creates a SidebarEmbedInfo from an InlineImageMatch (image embed)
-    static func fromImageMatch(_ match: InlineImageMatch, relativeTo noteURL: URL?) -> SidebarEmbedInfo {
-        let resolved = resolvedSidebarImageURL(for: match.source, relativeTo: noteURL)
-        return SidebarEmbedInfo(
-            id: match.id,
-            type: .image,
-            title: nil,
-            caption: match.caption.isEmpty ? nil : match.caption,
-            content: nil,
-            source: match.source,
-            resolvedURL: resolved,
-            isUnresolved: resolved == nil,
-            range: match.range
-        )
-    }
-}
-
-/// Resolves an image source string to a URL for sidebar display
-func resolvedSidebarImageURL(for source: String, relativeTo noteURL: URL?) -> URL? {
-    let cleanedSource = source.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !cleanedSource.isEmpty else { return nil }
-
-    // Handle web URLs
-    if cleanedSource.hasPrefix("http://") || cleanedSource.hasPrefix("https://") {
-        return URL(string: cleanedSource)
-    }
-
-    // Handle file:// URLs
-    if cleanedSource.hasPrefix("file://") {
-        return URL(string: cleanedSource)
-    }
-
-    // Handle absolute paths
-    if cleanedSource.hasPrefix("/") {
-        return URL(fileURLWithPath: cleanedSource)
-    }
-
-    // Handle relative paths
-    guard let noteURL = noteURL else { return nil }
-    let baseURL = noteURL.deletingLastPathComponent()
-    return URL(fileURLWithPath: cleanedSource, relativeTo: baseURL).standardizedFileURL
-}
-
-// MARK: - Embedded Notes Side Panel
-
-struct EmbeddedNotesPanel: NSViewRepresentable {
-    let notes: [SidebarEmbedInfo]
-    let allFiles: [URL]
-    let selectedEmbedID: String?
-    let onOpenFile: (URL, Bool) -> Void // (url, openInNewTab)
-    let onScrollToEmbed: ((NSRange) -> Void)?
-
-    func makeNSView(context: Context) -> NSScrollView {
-        let scrollView = NSScrollView()
-        scrollView.hasVerticalScroller = true
-        scrollView.autohidesScrollers = true
-        scrollView.borderType = .noBorder
-        scrollView.drawsBackground = true
-        scrollView.backgroundColor = SynapseTheme.editorBackground
-
-        let documentView = FlippedNSView()
-        documentView.autoresizingMask = [.width]
-        documentView.wantsLayer = true
-        documentView.layer?.backgroundColor = SynapseTheme.editorBackground.cgColor
-        scrollView.documentView = documentView
-
-        return scrollView
-    }
-
-    func updateNSView(_ scrollView: NSScrollView, context: Context) {
-        guard let documentView = scrollView.documentView else { return }
-
-        scrollView.drawsBackground = true
-        scrollView.backgroundColor = SynapseTheme.editorBackground
-        scrollView.contentView.backgroundColor = SynapseTheme.editorBackground
-        scrollView.documentView?.wantsLayer = true
-        scrollView.documentView?.layer?.backgroundColor = SynapseTheme.editorBackground.cgColor
-
-        let width: CGFloat = 304 // 320 - 16 padding
-        let spacing: CGFloat = 12
-        var currentY: CGFloat = 8
-        var selectedView: NSView?
-        var selectedViewY: CGFloat = 0
-
-        // Track which embed IDs we've processed
-        var processedIDs = Set<String>()
-
-        for embed in notes {
-            processedIDs.insert(embed.id)
-            let isSelected = embed.id == selectedEmbedID
-
-            // Find existing view for this embed ID
-            let existingView = documentView.subviews.first { $0.identifier?.rawValue == embed.id }
-
-            switch embed.type {
-            case .note:
-                let embedView: EmbeddedNoteView
-                if let existing = existingView as? EmbeddedNoteView {
-                    embedView = existing
-                } else {
-                    embedView = EmbeddedNoteView()
-                    embedView.identifier = NSUserInterfaceItemIdentifier(embed.id)
-                    embedView.onOpenNote = { url, openInNewTab in
-                        onOpenFile(url, openInNewTab)
-                    }
-                    documentView.addSubview(embedView)
-                }
-
-                embedView.configure(
-                    noteName: embed.title ?? "Note",
-                    content: embed.content,
-                    noteURL: embed.resolvedURL,
-                    isUnresolved: embed.isUnresolved
-                )
-
-                // Calculate height
-                let preferredSize = embedView.preferredSize(for: embed.content)
-                let height = min(preferredSize.height, 400)
-
-                embedView.frame = NSRect(x: 0, y: currentY, width: width, height: height)
-
-                if isSelected {
-                    selectedView = embedView
-                    selectedViewY = currentY
-                }
-
-                currentY += height + spacing
-
-            case .image:
-                let imageView: EmbeddedImageView
-                if let existing = existingView as? EmbeddedImageView {
-                    imageView = existing
-                } else {
-                    imageView = EmbeddedImageView()
-                    imageView.identifier = NSUserInterfaceItemIdentifier(embed.id)
-                    imageView.onScrollToMarkdown = { [range = embed.range] in
-                        onScrollToEmbed?(range)
-                    }
-                    documentView.addSubview(imageView)
-                }
-
-                imageView.configure(
-                    caption: embed.caption,
-                    imageURL: embed.resolvedURL,
-                    isUnresolved: embed.isUnresolved,
-                    isSelected: isSelected
-                )
-
-                let height: CGFloat = embed.caption != nil ? 246 : 228
-                imageView.frame = NSRect(x: 0, y: currentY, width: width, height: height)
-
-                if isSelected {
-                    selectedView = imageView
-                    selectedViewY = currentY
-                }
-
-                currentY += height + spacing
-            }
-        }
-
-        // Remove views that are no longer needed
-        documentView.subviews.forEach { view in
-            if let id = view.identifier?.rawValue, !processedIDs.contains(id) {
-                view.removeFromSuperview()
-            }
-        }
-
-        // Set document view size
-        let totalHeight = max(currentY - spacing + 8, scrollView.bounds.height)
-        documentView.frame = NSRect(x: 0, y: 0, width: width, height: totalHeight)
-
-        // Scroll selected view into view
-        if let selectedView = selectedView {
-            let visibleRect = NSRect(
-                x: 0,
-                y: selectedViewY,
-                width: width,
-                height: selectedView.frame.height
-            )
-            scrollView.contentView.scrollToVisible(visibleRect)
-        }
-    }
-}
-
-// NSView subclass with flipped coordinate system so (0,0) is at top-left
-final class FlippedNSView: NSView {
-    override var isFlipped: Bool { true }
-}
-
-final class EmbeddedNoteView: NSView {
-    private let contentScrollView = NSScrollView()
-    private let contentTextView = NSTextView()
-    private let titleField = NSTextField(labelWithString: "")
-    private let borderView = NSView()
-    private let openButton = NSButton()
-    private var targetURL: URL?
-    var onOpenNote: ((URL, Bool) -> Void)? // (url, openInNewTab)
-
-    // Fixed dimensions for the right-aligned panel
-    private let panelWidth: CGFloat = 280
-    private let maxPanelHeight: CGFloat = 400
-    private let minPanelHeight: CGFloat = 120
-
-    override init(frame frameRect: NSRect) {
-        super.init(frame: frameRect)
-        setup()
-    }
-
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        setup()
-    }
-
-    func configure(noteName: String, content: String?, noteURL: URL?, isUnresolved: Bool) {
-        targetURL = noteURL
-        titleField.stringValue = isUnresolved ? "Note not found: \(noteName)" : noteName
-
-        if isUnresolved {
-            contentTextView.string = ""
-            contentScrollView.isHidden = true
-            borderView.layer?.backgroundColor = SynapseTheme.nsPanelElevated.cgColor
-            borderView.layer?.borderColor = SynapseTheme.nsError.cgColor
-        } else if let content = content {
-            let styledContent = styleMarkdownContent(content, fontSize: 11)
-            contentTextView.textStorage?.setAttributedString(styledContent)
-            contentScrollView.isHidden = false
-            borderView.layer?.backgroundColor = SynapseTheme.nsPanelElevated.cgColor
-            borderView.layer?.borderColor = SynapseTheme.nsBorder.cgColor
-        }
-
-        openButton.isHidden = (noteURL == nil)
-        updateColors()
-    }
-
-    /// Re-applies all theme-dependent colors. Safe to call any time the theme changes.
-    func updateColors() {
-        borderView.layer?.backgroundColor = SynapseTheme.nsPanelElevated.cgColor
-        titleField.textColor = SynapseTheme.nsTextPrimary
-        contentTextView.backgroundColor = SynapseTheme.editorCodeBackground
-        contentTextView.textColor = SynapseTheme.nsTextSecondary
-        contentScrollView.backgroundColor = SynapseTheme.editorCodeBackground
-        // Re-style markdown content with the new theme colors
-        if let text = contentTextView.string.isEmpty ? nil : contentTextView.string {
-            let styledContent = styleMarkdownContent(text, fontSize: 11)
-            contentTextView.textStorage?.setAttributedString(styledContent)
-        }
-    }
-
-    override func layout() {
-        super.layout()
-        wantsLayer = true
-        layer?.cornerRadius = 6
-        layer?.masksToBounds = true
-
-        let padding: CGFloat = 12
-        let buttonHeight: CGFloat = 28
-        let titleHeight: CGFloat = 20
-        let spacing: CGFloat = 8
-
-        // Border view fills the entire frame
-        borderView.frame = bounds
-
-        // Title at top
-        titleField.frame = NSRect(
-            x: padding,
-            y: bounds.height - padding - titleHeight,
-            width: bounds.width - padding * 2,
-            height: titleHeight
-        )
-
-        // Open button at bottom
-        openButton.frame = NSRect(
-            x: bounds.width - padding - 80,
-            y: padding,
-            width: 80,
-            height: buttonHeight
-        )
-
-        // Content scroll view fills the middle area
-        if !contentScrollView.isHidden {
-            let contentY = buttonHeight + padding + spacing
-            let contentHeight = bounds.height - contentY - titleHeight - spacing * 2
-            contentScrollView.frame = NSRect(
-                x: padding,
-                y: contentY,
-                width: bounds.width - padding * 2,
-                height: max(0, contentHeight)
-            )
-        }
-    }
-
-    @objc private func openNote() {
-        guard let url = targetURL else { return }
-        // Check if Command key is held (for opening in new tab)
-        let openInNewTab = NSEvent.modifierFlags.contains(.command)
-        onOpenNote?(url, openInNewTab)
-    }
-
-    private func setup() {
-        // Border view
-        borderView.wantsLayer = true
-        borderView.layer?.cornerRadius = 6
-        borderView.layer?.masksToBounds = true
-        borderView.layer?.borderWidth = 1
-        borderView.layer?.backgroundColor = SynapseTheme.nsPanelElevated.cgColor
-        borderView.layer?.borderColor = SynapseTheme.nsBorder.cgColor
-        borderView.autoresizingMask = [.width, .height]
-        addSubview(borderView)
-
-        // Title field
-        titleField.font = .systemFont(ofSize: 13, weight: .semibold)
-        titleField.textColor = SynapseTheme.nsTextPrimary
-        titleField.lineBreakMode = .byTruncatingTail
-        addSubview(titleField)
-
-        // Content text view (read-only)
-        contentTextView.isEditable = false
-        contentTextView.isSelectable = true
-        contentTextView.isRichText = false
-        contentTextView.backgroundColor = SynapseTheme.editorCodeBackground
-        contentTextView.textContainerInset = NSSize(width: 8, height: 8)
-        contentTextView.font = .systemFont(ofSize: 11)
-        contentTextView.textColor = SynapseTheme.nsTextSecondary
-
-        // Content scroll view
-        contentScrollView.documentView = contentTextView
-        contentScrollView.hasVerticalScroller = true
-        contentScrollView.autohidesScrollers = true
-        contentScrollView.borderType = .bezelBorder
-        contentScrollView.backgroundColor = SynapseTheme.editorCodeBackground
-        contentScrollView.isHidden = true
-        addSubview(contentScrollView)
-
-        // Open button
-        openButton.title = "Open"
-        openButton.target = self
-        openButton.action = #selector(openNote)
-        openButton.bezelStyle = .rounded
-        openButton.font = .systemFont(ofSize: 11, weight: .medium)
-        addSubview(openButton)
-    }
-
-    // Return the preferred size for this panel
-    func preferredSize(for content: String?) -> NSSize {
-        let padding: CGFloat = 12
-        let buttonHeight: CGFloat = 28
-        let titleHeight: CGFloat = 20
-        let spacing: CGFloat = 8
-
-        if content == nil {
-            // Unresolved: just title + button
-            return NSSize(width: panelWidth, height: minPanelHeight)
-        }
-
-        // Calculate content height based on text
-        let textStorage = NSTextStorage(string: content!)
-        let layoutManager = NSLayoutManager()
-        let textContainer = NSTextContainer(containerSize: NSSize(width: panelWidth - padding * 2 - 20, height: .greatestFiniteMagnitude))
-        textContainer.lineFragmentPadding = 0
-
-        layoutManager.addTextContainer(textContainer)
-        textStorage.addLayoutManager(layoutManager)
-        layoutManager.ensureLayout(for: textContainer)
-
-        let contentHeight = layoutManager.usedRect(for: textContainer).height + 16 // +16 for insets
-        let totalHeight = padding + buttonHeight + spacing + min(contentHeight, 300) + spacing + titleHeight + padding
-
-        return NSSize(width: panelWidth, height: min(max(totalHeight, minPanelHeight), maxPanelHeight))
-    }
-}
-
-// MARK: - Embedded Image View
-
-final class EmbeddedImageView: NSView {
-    private let imageView = NSImageView()
-    private let captionField = NSTextField(labelWithString: "")
-    private let borderView = NSView()
-    private let previewBackgroundView = NSView()
-    private let openButton = NSButton()
-    private var targetURL: URL?
-    private var isSelected: Bool = false
-    var onScrollToMarkdown: (() -> Void)?
-
-    override init(frame frameRect: NSRect) {
-        super.init(frame: frameRect)
-        setup()
-    }
-
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        setup()
-    }
-
-    func configure(caption: String?, imageURL: URL?, isUnresolved: Bool, isSelected: Bool = false) {
-        targetURL = imageURL
-        self.isSelected = isSelected
-        openButton.isHidden = imageURL == nil || isUnresolved
-
-        if isUnresolved {
-            captionField.stringValue = caption ?? "Image not found"
-            imageView.image = nil
-        } else {
-            captionField.stringValue = caption ?? ""
-            // Load image asynchronously
-            if let imageURL = imageURL {
-                loadImage(from: imageURL)
-            }
-        }
-
-        captionField.isHidden = (caption == nil || caption?.isEmpty == true)
-
-        // Update border color based on selection state
-        updateBorderAppearance()
-        updateColors()
-    }
-
-    private func updateBorderAppearance() {
-        borderView.layer?.borderWidth = isSelected ? 3 : 1
-        borderView.layer?.borderColor = isSelected
-            ? NSColor(SynapseTheme.accent).cgColor
-            : NSColor(SynapseTheme.border).cgColor
-    }
-
-    /// Re-applies all theme-dependent colors. Safe to call any time the theme changes.
-    func updateColors() {
-        borderView.layer?.backgroundColor = SynapseTheme.nsPanelElevated.cgColor
-        previewBackgroundView.layer?.backgroundColor = SynapseTheme.editorCodeBackground.cgColor
-        updateBorderAppearance()
-        captionField.textColor = NSColor(SynapseTheme.textSecondary)
-    }
-
-    private func loadImage(from url: URL) {
-        // Load image in background
-        DispatchQueue.global(qos: .userInitiated).async { [weak self] in
-            guard let image = NSImage(contentsOf: url) else { return }
-            DispatchQueue.main.async {
-                self?.imageView.image = image
-                self?.needsLayout = true
-            }
-        }
-    }
-
-    override func layout() {
-        super.layout()
-
-        wantsLayer = true
-        layer?.cornerRadius = 6
-        layer?.masksToBounds = true
-
-        // Update border frame and appearance
-        borderView.frame = bounds
-        updateBorderAppearance()
-
-        let padding: CGFloat = 12
-        let spacing: CGFloat = 10
-        let buttonHeight: CGFloat = openButton.isHidden ? 0 : 24
-        let captionHeight: CGFloat = captionField.isHidden ? 0 : 20
-
-        let buttonY = padding
-        let previewBottom = buttonY + buttonHeight + (openButton.isHidden ? 0 : spacing)
-        let previewTop = bounds.height - padding - captionHeight - (captionField.isHidden ? 0 : spacing)
-        let previewRect = NSRect(
-            x: padding,
-            y: previewBottom,
-            width: bounds.width - padding * 2,
-            height: max(120, previewTop - previewBottom)
-        )
-
-        previewBackgroundView.frame = previewRect
-
-        let contentRect = previewRect.insetBy(dx: 8, dy: 8)
-        if let image = imageView.image, image.size.width > 0, image.size.height > 0 {
-            let widthRatio = contentRect.width / image.size.width
-            let heightRatio = contentRect.height / image.size.height
-            let scale = min(widthRatio, heightRatio)
-            let drawSize = NSSize(width: image.size.width * scale, height: image.size.height * scale)
-            imageView.frame = NSRect(
-                x: round(contentRect.midX - drawSize.width / 2),
-                y: round(contentRect.midY - drawSize.height / 2),
-                width: round(drawSize.width),
-                height: round(drawSize.height)
-            )
-        } else {
-            imageView.frame = contentRect
-        }
-        imageView.imageScaling = .scaleProportionallyUpOrDown
-        imageView.contentTintColor = nil
-
-        // Caption label
-        if !captionField.isHidden {
-            captionField.frame = NSRect(
-                x: padding,
-                y: bounds.height - padding - captionHeight,
-                width: bounds.width - padding * 2,
-                height: captionHeight
-            )
-            captionField.font = .monospacedSystemFont(ofSize: 12, weight: .semibold)
-            captionField.textColor = NSColor(SynapseTheme.textSecondary)
-            captionField.lineBreakMode = .byTruncatingMiddle
-            captionField.alignment = .left
-        }
-
-        let buttonWidth = min(124, bounds.width - padding * 2)
-        openButton.frame = NSRect(
-            x: round((bounds.width - buttonWidth) / 2),
-            y: padding,
-            width: buttonWidth,
-            height: buttonHeight
-        )
-        openButton.bezelStyle = .rounded
-        openButton.font = .systemFont(ofSize: 11, weight: .semibold)
-    }
-
-    private var imageViewerController: ImageViewerWindowController?
-
-    @objc private func openImage() {
-        guard let targetURL = targetURL else { return }
-
-        let viewer = ImageViewerWindowController(imageURL: targetURL, caption: captionField.stringValue.isEmpty ? nil : captionField.stringValue)
-        imageViewerController = viewer // retain strongly so it isn't deallocated before image loads
-        viewer.showFullScreen()
-    }
-
-    private func setup() {
-        // Setup border view layer properties
-        borderView.wantsLayer = true
-        borderView.layer?.cornerRadius = 6
-        borderView.layer?.masksToBounds = true
-        borderView.layer?.backgroundColor = SynapseTheme.nsPanelElevated.cgColor
-
-        addSubview(borderView)
-        previewBackgroundView.wantsLayer = true
-        previewBackgroundView.layer?.cornerRadius = 8
-        previewBackgroundView.layer?.masksToBounds = true
-        previewBackgroundView.layer?.backgroundColor = SynapseTheme.editorCodeBackground.cgColor
-        previewBackgroundView.layer?.borderWidth = 0
-        addSubview(previewBackgroundView)
-        addSubview(imageView)
-        addSubview(captionField)
-
-        openButton.title = "Open"
-        openButton.bezelStyle = .rounded
-        openButton.target = self
-        openButton.action = #selector(openImage)
-        addSubview(openButton)
-
-        // Click on image thumbnail scrolls editor to the markdown
-        let click = NSClickGestureRecognizer(target: self, action: #selector(thumbnailClicked))
-        imageView.addGestureRecognizer(click)
-
-        // Initial border appearance
-        updateBorderAppearance()
-    }
-
-    @objc private func thumbnailClicked() {
-        onScrollToMarkdown?()
-    }
-}
-
-// MARK: - Full Screen Image Viewer
-
-/// A full-screen window for viewing images with zoom and pan support
-final class ImageViewerWindowController: NSWindowController {
-    private let imageView = NSImageView()
-    private let imageContainerView = NSView()
-    private var imageURL: URL?
-    private var localMonitor: Any?
-    private var scrollMonitor: Any?
-    private var scrollView: NSScrollView!
-    private var currentZoom: CGFloat = 1.0
-    private var minZoom: CGFloat = 0.1
-    private var maxZoom: CGFloat = 5.0
-    private var imageSize: NSSize = .zero
-    private var imageWidthConstraint: NSLayoutConstraint?
-    private var imageHeightConstraint: NSLayoutConstraint?
-    private var gestureStartZoom: CGFloat = 1.0
-
-    init(imageURL: URL, caption: String?) {
-        let window = NSWindow(
-            contentRect: NSScreen.main?.frame ?? NSRect(x: 0, y: 0, width: 800, height: 600),
-            styleMask: [.titled, .closable, .fullSizeContentView, .resizable],
-            backing: .buffered,
-            defer: false
-        )
-        window.title = caption ?? imageURL.lastPathComponent
-        window.titlebarAppearsTransparent = false
-        window.backgroundColor = .black
-        window.isOpaque = true
-        window.hasShadow = true
-
-        super.init(window: window)
-
-        self.imageURL = imageURL
-        setupContentView()
-        setupImageView()
-        setupCloseButton()
-        setupEscapeHandler()
-        loadImage()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    deinit {
-        if let monitor = localMonitor {
-            NSEvent.removeMonitor(monitor)
-        }
-        if let monitor = scrollMonitor {
-            NSEvent.removeMonitor(monitor)
-        }
-    }
-
-    private func setupContentView() {
-        guard let window = window else { return }
-
-        scrollView = NSScrollView(frame: window.contentView?.bounds ?? .zero)
-        scrollView.autoresizingMask = [.width, .height]
-        scrollView.hasVerticalScroller = false
-        scrollView.hasHorizontalScroller = false
-        scrollView.autohidesScrollers = true
-        scrollView.allowsMagnification = false
-        scrollView.backgroundColor = .black
-        scrollView.drawsBackground = true
-
-        window.contentView = scrollView
-    }
-
-    private func setupImageView() {
-        imageContainerView.wantsLayer = true
-        imageContainerView.layer?.backgroundColor = NSColor.black.cgColor
-        imageContainerView.frame = scrollView.bounds
-
-        imageView.imageScaling = .scaleAxesIndependently
-        imageView.imageAlignment = .alignCenter
-        imageView.wantsLayer = true
-        imageView.layer?.backgroundColor = NSColor.clear.cgColor
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-
-        imageContainerView.addSubview(imageView)
-        imageWidthConstraint = imageView.widthAnchor.constraint(equalToConstant: 100)
-        imageHeightConstraint = imageView.heightAnchor.constraint(equalToConstant: 100)
-
-        NSLayoutConstraint.activate([
-            imageView.centerXAnchor.constraint(equalTo: imageContainerView.centerXAnchor),
-            imageView.centerYAnchor.constraint(equalTo: imageContainerView.centerYAnchor),
-            imageWidthConstraint!,
-            imageHeightConstraint!
-        ])
-
-        scrollView.documentView = imageContainerView
-
-        setupGestureRecognizers()
-        setupScrollWheelZoom()
-    }
-
-    private func setupScrollWheelZoom() {
-        scrollMonitor = NSEvent.addLocalMonitorForEvents(matching: .scrollWheel) { [weak self] event in
-            guard let self = self, let window = self.window else { return event }
-
-            if NSApp.keyWindow == window && event.modifierFlags.contains(.control) {
-                let delta = event.scrollingDeltaY
-                let zoomFactor = pow(1.01, delta * 0.35)
-                let newZoom = self.currentZoom * zoomFactor
-                self.setZoom(newZoom, animated: false)
-                return nil
-            }
-            return event
-        }
-    }
-
-    private func setupGestureRecognizers() {
-        let doubleClickGesture = NSClickGestureRecognizer(target: self, action: #selector(handleDoubleClick))
-        doubleClickGesture.numberOfClicksRequired = 2
-        imageView.addGestureRecognizer(doubleClickGesture)
-
-        let magnifyGesture = NSMagnificationGestureRecognizer(target: self, action: #selector(handleMagnify(_:)))
-        imageView.addGestureRecognizer(magnifyGesture)
-    }
-
-    @objc private func handleDoubleClick() {
-        // Toggle between fit-to-screen and 100% zoom
-        if currentZoom != 1.0 {
-            setZoom(1.0, animated: true)
-        } else {
-            fitImageToScreen()
-        }
-    }
-
-    @objc private func handleMagnify(_ gesture: NSMagnificationGestureRecognizer) {
-        switch gesture.state {
-        case .began:
-            gestureStartZoom = currentZoom
-        case .changed:
-            let newZoom = gestureStartZoom * (1 + gesture.magnification)
-            setZoom(newZoom, animated: false)
-        default:
-            break
-        }
-    }
-
-    private func setZoom(_ zoom: CGFloat, animated: Bool) {
-        let clampedZoom = max(minZoom, min(maxZoom, zoom))
-        currentZoom = clampedZoom
-
-        let applyLayout = { self.layoutImage(centerViewport: true) }
-        if animated {
-            NSAnimationContext.runAnimationGroup { context in
-                context.duration = 0.15
-                applyLayout()
-            }
-        } else {
-            applyLayout()
-        }
-    }
-
-    private func fitImageToScreen() {
-        guard imageSize != .zero, let window = window else { return }
-
-        let visibleFrame = window.contentView?.bounds ?? window.frame
-        let titleBarHeight: CGFloat = 28
-        let availableHeight = visibleFrame.height - titleBarHeight - 40
-        let availableWidth = visibleFrame.width - 40
-
-        let widthRatio = availableWidth / imageSize.width
-        let heightRatio = availableHeight / imageSize.height
-        currentZoom = min(widthRatio, heightRatio, 1.0)
-        layoutImage(centerViewport: true)
-    }
-
-    private func layoutImage(centerViewport: Bool) {
-        guard imageSize != .zero else { return }
-
-        let visibleSize = scrollView.contentView.bounds.size
-        let scaledSize = NSSize(width: imageSize.width * currentZoom, height: imageSize.height * currentZoom)
-        let containerSize = NSSize(
-            width: max(visibleSize.width, scaledSize.width),
-            height: max(visibleSize.height, scaledSize.height)
-        )
-
-        imageContainerView.frame = NSRect(origin: .zero, size: containerSize)
-        imageWidthConstraint?.constant = scaledSize.width
-        imageHeightConstraint?.constant = scaledSize.height
-        imageContainerView.layoutSubtreeIfNeeded()
-
-        if centerViewport {
-            let centeredOrigin = NSPoint(
-                x: max(0, (containerSize.width - visibleSize.width) / 2),
-                y: max(0, (containerSize.height - visibleSize.height) / 2)
-            )
-            scrollView.contentView.scroll(to: centeredOrigin)
-            scrollView.reflectScrolledClipView(scrollView.contentView)
-        }
-    }
-
-    private func setupCloseButton() {
-        // Native window close button (traffic light) is sufficient
-    }
-
-    private func setupEscapeHandler() {
-        localMonitor = NSEvent.addLocalMonitorForEvents(matching: .keyDown) { [weak self] event in
-            guard let self = self, let window = self.window else { return event }
-
-            if NSApp.keyWindow == window && event.keyCode == 53 {
-                self.closeWindow()
-                return nil
-            }
-            return event
-        }
-    }
-
-    private func loadImage() {
-        guard let imageURL = imageURL else { return }
-
-        // Handle remote URLs (http/https)
-        if imageURL.scheme?.lowercased() == "http" || imageURL.scheme?.lowercased() == "https" {
-            downloadRemoteImage(from: imageURL)
-            return
-        }
-
-        // Handle local file URLs
-        let fileManager = FileManager.default
-        if !fileManager.fileExists(atPath: imageURL.path) {
-            print("Image file does not exist at: \(imageURL.path)")
-            return
-        }
-
-        DispatchQueue.global(qos: .userInitiated).async { [weak self] in
-            guard let image = NSImage(contentsOf: imageURL) else {
-                print("Failed to load image from: \(imageURL.path)")
-                return
-            }
-
-            DispatchQueue.main.async {
-                self?.imageSize = image.size
-                self?.imageView.image = image
-                self?.updateImageViewSize()
-                self?.fitImageToScreen()
-                print("Image loaded successfully: \(image.size.width)x\(image.size.height)")
-            }
-        }
-    }
-
-    private func downloadRemoteImage(from url: URL) {
-        print("Downloading remote image from: \(url.absoluteString)")
-
-        let task = URLSession.shared.dataTask(with: url) { [weak self] data, response, error in
-            if let error = error {
-                print("Failed to download image: \(error.localizedDescription)")
-                return
-            }
-
-            guard let data = data, let image = NSImage(data: data) else {
-                print("Failed to create image from downloaded data")
-                return
-            }
-
-            DispatchQueue.main.async {
-                self?.imageSize = image.size
-                self?.imageView.image = image
-                self?.updateImageViewSize()
-                self?.fitImageToScreen()
-                print("Remote image loaded successfully: \(image.size.width)x\(image.size.height)")
-            }
-        }
-
-        task.resume()
-    }
-
-    private func updateImageViewSize() {
-        layoutImage(centerViewport: true)
-    }
-
-    @objc private func closeWindow() {
-        window?.close()
-    }
-
-    func showFullScreen() {
-        window?.center()
-        window?.makeKeyAndOrderFront(nil)
-
-        // Make it nearly full screen but keep title bar
-        if let screen = NSScreen.main {
-            let screenFrame = screen.visibleFrame
-            let padding: CGFloat = 40
-            let newFrame = NSRect(
-                x: screenFrame.origin.x + padding,
-                y: screenFrame.origin.y + padding,
-                width: screenFrame.width - (padding * 2),
-                height: screenFrame.height - (padding * 2)
-            )
-            window?.setFrame(newFrame, display: true, animate: true)
-        }
-    }
-}
-
-final class YouTubePreviewView: NSView {
-    private let thumbnailView = NSImageView()
-    private let overlay = NSView()
-    private let playIcon = NSImageView()
-    private let titleField = NSTextField(labelWithString: "")
-    private let subtitleField = NSTextField(labelWithString: "")
-    private let actionButton = NSButton()
-    private var targetURL: URL?
-
-    override init(frame frameRect: NSRect) {
-        super.init(frame: frameRect)
-        setup()
-    }
-
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        setup()
-    }
-
-    override func layout() {
-        super.layout()
-        wantsLayer = true
-        layer?.cornerRadius = 4
-        layer?.masksToBounds = true
-        layer?.borderWidth = 1
-        layer?.borderColor = NSColor(SynapseTheme.border).cgColor
-        layer?.backgroundColor = SynapseTheme.editorCodeBackground.cgColor
-
-        thumbnailView.frame = bounds
-        overlay.frame = bounds
-        actionButton.frame = bounds
-
-        let iconSize: CGFloat = 54
-        playIcon.frame = NSRect(x: 20, y: bounds.midY - iconSize / 2, width: iconSize, height: iconSize)
-
-        let textX = playIcon.frame.maxX + 18
-        let textWidth = max(160, bounds.width - textX - 20)
-        titleField.frame = NSRect(x: textX, y: bounds.midY + 2, width: textWidth, height: 28)
-        subtitleField.frame = NSRect(x: textX, y: bounds.midY - 28, width: textWidth, height: 44)
-    }
-
-    @objc private func openVideo() {
-        guard let targetURL else { return }
-        NSWorkspace.shared.open(targetURL)
-    }
-
-    private func setup() {
-        thumbnailView.imageScaling = .scaleProportionallyUpOrDown
-        thumbnailView.imageAlignment = .alignCenter
-        thumbnailView.autoresizingMask = [.width, .height]
-        addSubview(thumbnailView)
-
-        overlay.wantsLayer = true
-        overlay.layer?.backgroundColor = NSColor.black.withAlphaComponent(0.35).cgColor
-        overlay.autoresizingMask = [.width, .height]
-        addSubview(overlay)
-
-        if let image = NSImage(systemSymbolName: "play.circle.fill", accessibilityDescription: nil) {
-            playIcon.image = image
-        }
-        playIcon.contentTintColor = .white
-        addSubview(playIcon)
-
-        titleField.font = .systemFont(ofSize: 20, weight: .bold)
-        titleField.textColor = NSColor(SynapseTheme.textPrimary)
-        titleField.lineBreakMode = .byTruncatingTail
-        addSubview(titleField)
-
-        subtitleField.font = .systemFont(ofSize: 12, weight: .medium)
-        subtitleField.textColor = NSColor(SynapseTheme.textSecondary)
-        subtitleField.lineBreakMode = .byTruncatingMiddle
-        addSubview(subtitleField)
-
-        actionButton.isBordered = false
-        actionButton.title = ""
-        actionButton.target = self
-        actionButton.action = #selector(openVideo)
-        actionButton.autoresizingMask = [.width, .height]
-        addSubview(actionButton)
-    }
-}
-
-// MARK: - Completion popover
-
-class CompletionViewController: NSViewController {
-    var onSelect: ((URL) -> Void)?
-    private let searchField = NSSearchField()
-    private let tableView = NSTableView()
-    private let scrollView = NSScrollView()
-    private var allFiles: [URL] = []
-    private var filteredFiles: [URL] = []
-
-    override func loadView() {
-        view = NSView(frame: NSRect(x: 0, y: 0, width: 420, height: 260))
-
-        searchField.placeholderString = "Search files..."
-        searchField.sendsSearchStringImmediately = true
-        searchField.target = self
-        searchField.action = #selector(searchChanged)
-        searchField.delegate = self
-        searchField.font = .systemFont(ofSize: 12)
-
-        let col = NSTableColumn(identifier: NSUserInterfaceItemIdentifier("name"))
-        col.isEditable = false
-        tableView.addTableColumn(col)
-        tableView.headerView = nil
-        tableView.rowHeight = 26
-        tableView.dataSource = self
-        tableView.delegate = self
-        tableView.doubleAction = #selector(selectItem)
-        tableView.target = self
-        tableView.allowsEmptySelection = false
-
-        scrollView.documentView = tableView
-        scrollView.hasVerticalScroller = true
-        scrollView.autohidesScrollers = true
-
-        let container = NSView()
-        container.translatesAutoresizingMaskIntoConstraints = false
-        searchField.translatesAutoresizingMaskIntoConstraints = false
-        scrollView.translatesAutoresizingMaskIntoConstraints = false
-        container.addSubview(searchField)
-        container.addSubview(scrollView)
-
-        NSLayoutConstraint.activate([
-            searchField.topAnchor.constraint(equalTo: container.topAnchor, constant: 8),
-            searchField.leadingAnchor.constraint(equalTo: container.leadingAnchor, constant: 8),
-            searchField.trailingAnchor.constraint(equalTo: container.trailingAnchor, constant: -8),
-            scrollView.topAnchor.constraint(equalTo: searchField.bottomAnchor, constant: 6),
-            scrollView.leadingAnchor.constraint(equalTo: container.leadingAnchor),
-            scrollView.trailingAnchor.constraint(equalTo: container.trailingAnchor),
-            scrollView.bottomAnchor.constraint(equalTo: container.bottomAnchor),
-        ])
-        view.addSubview(container)
-        NSLayoutConstraint.activate([
-            container.topAnchor.constraint(equalTo: view.topAnchor),
-            container.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            container.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            container.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            scrollView.heightAnchor.constraint(greaterThanOrEqualToConstant: 160),
-        ])
-    }
-
-    @objc private func searchChanged() {
-        applyFilter()
-    }
-
-    private func normalize(_ value: String) -> String {
-        value
-            .folding(options: [.diacriticInsensitive, .caseInsensitive], locale: .current)
-            .components(separatedBy: .newlines).joined()
-            .lowercased()
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-    }
-
-    private func scoreFile(_ url: URL, query: String) -> Int? {
-        if query.isEmpty { return 1 }
-        let name = normalize(url.deletingPathExtension().lastPathComponent)
-        let path = normalize(url.path)
-        if let range = name.range(of: query) {
-            let offset = name.distance(from: name.startIndex, to: range.lowerBound)
-            return 400 - min(offset, 300)
-        }
-        if let range = path.range(of: query) {
-            let offset = path.distance(from: path.startIndex, to: range.lowerBound)
-            return 200 - min(offset, 180)
-        }
-        return nil
-    }
-
-    private func applyFilter() {
-        let query = normalize(searchField.stringValue)
-        filteredFiles = allFiles
-            .compactMap { url -> (URL, Int)? in
-                guard let score = scoreFile(url, query: query) else { return nil }
-                return (url, score)
-            }
-            .sorted { $0.1 > $1.1 }
-            .map { $0.0 }
-        tableView.reloadData()
-        if !filteredFiles.isEmpty {
-            tableView.selectRowIndexes([0], byExtendingSelection: false)
-        }
-    }
-
-    @objc func selectItem() {
-        let row = tableView.selectedRow
-        guard row >= 0, row < filteredFiles.count else { return }
-        onSelect?(filteredFiles[row])
-    }
-
-    func selectCurrentItem() { selectItem() }
-
-    func moveSelection(by delta: Int) {
-        guard !filteredFiles.isEmpty else { return }
-        let current = max(0, tableView.selectedRow)
-        let next = max(0, min(filteredFiles.count - 1, current + delta))
-        tableView.selectRowIndexes([next], byExtendingSelection: false)
-        tableView.scrollRowToVisible(next)
-    }
-}
-
-extension CompletionViewController: NSTableViewDataSource, NSTableViewDelegate {
-    func numberOfRows(in tableView: NSTableView) -> Int { filteredFiles.count }
-
-    func tableView(_ tableView: NSTableView, viewFor tableColumn: NSTableColumn?, row: Int) -> NSView? {
-        let name = filteredFiles[row].deletingPathExtension().lastPathComponent
-        let cell = NSTextField(labelWithString: name)
-        cell.font = .systemFont(ofSize: 13)
-        cell.lineBreakMode = .byTruncatingMiddle
-        return cell
-    }
-
-    func tableViewSelectionDidChange(_ notification: Notification) {}
-}
-
-extension CompletionViewController: NSSearchFieldDelegate, NSControlTextEditingDelegate {
-    func control(_ control: NSControl, textView: NSTextView, doCommandBy commandSelector: Selector) -> Bool {
-        switch commandSelector {
-        case #selector(NSResponder.moveUp(_:)):
-            moveSelection(by: -1)
-            return true
-        case #selector(NSResponder.moveDown(_:)):
-            moveSelection(by: 1)
-            return true
-        case #selector(NSResponder.insertNewline(_:)):
-            selectCurrentItem()
-            return true
-        default:
-            return false
-        }
-    }
-}
-
-// MARK: - Code Block Copy Button
-
-private enum CodeBlockCopyButtonAssociatedKeys {
-    static var buttons: UInt8 = 0
-}
-
-final class CodeBlockCopyButton: NSButton {
-    var codeContent: String = ""
-
-    override init(frame frameRect: NSRect) {
-        super.init(frame: frameRect)
-        isBordered = false
-        bezelStyle = .inline
-        wantsLayer = true
-        layer?.cornerRadius = 4
-        layer?.backgroundColor = NSColor.windowBackgroundColor.withAlphaComponent(0.8).cgColor
-        image = NSImage(systemSymbolName: "doc.on.doc", accessibilityDescription: "Copy code")
-        imageScaling = .scaleProportionallyDown
-        contentTintColor = NSColor.secondaryLabelColor
-        toolTip = "Copy code"
-        target = self
-        action = #selector(handleClick)
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func resetCursorRects() {
-        discardCursorRects()
-        addCursorRect(bounds, cursor: .pointingHand)
-    }
-
-    override func acceptsFirstMouse(for event: NSEvent?) -> Bool {
-        true
-    }
-
-    @objc private func handleClick() {
-        let pasteboard = NSPasteboard.general
-        pasteboard.clearContents()
-        pasteboard.setString(codeContent, forType: .string)
-
-        contentTintColor = NSColor.systemGreen
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) { [weak self] in
-            self?.contentTintColor = NSColor.secondaryLabelColor
-        }
-    }
-}
-
-/// Represents a detected code block in markdown
-struct CodeBlockMatch: Equatable {
-    let id: String
-    let range: NSRange
-    let content: String
-    let language: String?
-
-    static func == (lhs: CodeBlockMatch, rhs: CodeBlockMatch) -> Bool {
-        lhs.id == rhs.id &&
-        lhs.range == rhs.range &&
-        lhs.content == rhs.content &&
-        lhs.language == rhs.language
-    }
-}
-
-extension LinkAwareTextView {
-
-    var codeBlockCopyButtons: [String: NSButton] {
-        get {
-            (objc_getAssociatedObject(self, &CodeBlockCopyButtonAssociatedKeys.buttons) as? [String: NSButton]) ?? [:]
-        }
-        set {
-            objc_setAssociatedObject(self, &CodeBlockCopyButtonAssociatedKeys.buttons, newValue, .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
-        }
-    }
-
-    /// Regex pattern to detect code blocks: ```optional_language\ncode\n```
-    /// Only matches opening ``` at the start of a line or string
-    private static let codeBlockRegex = try? NSRegularExpression(
-        pattern: "^[ \\t]{0,3}```([a-zA-Z0-9+-]*)[ \\t]*$",
-        options: [.anchorsMatchLines]
-    )
-
-    /// Find all code blocks in the current text
-    func codeBlockMatches() -> [CodeBlockMatch] {
-        guard let regex = Self.codeBlockRegex else { return [] }
-        let nsText = string as NSString
-        let fullRange = NSRange(location: 0, length: nsText.length)
-
-        let fenceMatches = regex.matches(in: string, options: [], range: fullRange)
-        var matches: [CodeBlockMatch] = []
-        var index = 0
-
-        while index + 1 < fenceMatches.count {
-            let openingMatch = fenceMatches[index]
-            let closingMatch = fenceMatches[index + 1]
-            let openingRange = openingMatch.range(at: 0)
-            let languageRange = openingMatch.range(at: 1)
-            let closingRange = closingMatch.range(at: 0)
-
-            let contentStart = openingRange.location + openingRange.length
-            let contentLength = closingRange.location - contentStart
-            guard contentLength >= 0 else {
-                index += 2
-                continue
-            }
-
-            let contentRange = NSRange(location: contentStart, length: contentLength)
-            var content = nsText.substring(with: contentRange)
-            if content.hasPrefix("\r\n") {
-                content.removeFirst(2)
-            } else if content.hasPrefix("\n") {
-                content.removeFirst()
-            }
-            if content.hasSuffix("\r\n") {
-                content.removeLast(2)
-            } else if content.hasSuffix("\n") {
-                content.removeLast()
-            }
-
-            let language = languageRange.length > 0 ? nsText.substring(with: languageRange) : nil
-            let fullRange = NSRange(location: openingRange.location, length: closingRange.location + closingRange.length - openingRange.location)
-            let id = "\(openingRange.location)-\(openingRange.length)"
-
-            matches.append(CodeBlockMatch(
-                id: id,
-                range: fullRange,
-                content: content,
-                language: language
-            ))
-
-            index += 2
-        }
-
-        return matches
-    }
-
-    /// Create and position copy buttons for all code blocks
-    func refreshCodeBlockCopyButtons() {
-        guard let layoutManager = layoutManager,
-              let textContainer = textContainer else { return }
-
-        layoutManager.ensureLayout(for: textContainer)
-
-        let matches = codeBlockMatches()
-        let activeKeys = Set(matches.map(\.id))
-
-        // Remove stale buttons
-        for key in Array(codeBlockCopyButtons.keys) where !activeKeys.contains(key) {
-            codeBlockCopyButtons[key]?.removeFromSuperview()
-            codeBlockCopyButtons.removeValue(forKey: key)
-        }
-        let buttonSize: CGFloat = 24
-        let buttonMargin: CGFloat = 8
-        let minBlockHeight = buttonSize + buttonMargin * 2
-
-        for match in matches {
-            // Get the rect of the code block
-            let glyphRange = layoutManager.glyphRange(forCharacterRange: match.range, actualCharacterRange: nil)
-            var codeBlockRect = layoutManager.boundingRect(forGlyphRange: glyphRange, in: textContainer)
-            codeBlockRect.origin.x += textContainerOrigin.x
-            codeBlockRect.origin.y += textContainerOrigin.y
-            // Guarantee enough height for the button even on very short blocks
-            if codeBlockRect.height < minBlockHeight {
-                codeBlockRect.size.height = minBlockHeight
-            }
-
-            // Position button at top-right corner
-            let buttonX = codeBlockRect.maxX - buttonSize - buttonMargin
-            let buttonY = codeBlockRect.minY + buttonMargin
-
-            let button: CodeBlockCopyButton
-            if let existing = codeBlockCopyButtons[match.id] {
-                guard let existingButton = existing as? CodeBlockCopyButton else {
-                    existing.removeFromSuperview()
-                    let replacementButton = createCopyButton(for: match)
-                    addSubview(replacementButton, positioned: .above, relativeTo: nil)
-                    codeBlockCopyButtons[match.id] = replacementButton
-                    replacementButton.frame = NSRect(x: buttonX, y: buttonY, width: buttonSize, height: buttonSize)
-                    continue
-                }
-                button = existingButton
-            } else {
-                button = createCopyButton(for: match)
-                addSubview(button, positioned: .above, relativeTo: nil)
-                codeBlockCopyButtons[match.id] = button
-            }
-
-            button.codeContent = match.content
-            button.frame = NSRect(x: buttonX, y: buttonY, width: buttonSize, height: buttonSize)
-        }
-    }
-
-    /// Create a copy button for a specific code block
-    private func createCopyButton(for match: CodeBlockMatch) -> CodeBlockCopyButton {
-        let button = CodeBlockCopyButton(frame: .zero)
-        button.identifier = NSUserInterfaceItemIdentifier(match.id)
-        button.codeContent = match.content
-        return button
-    }
-
-    /// Remove all code block copy buttons
-    func clearCodeBlockCopyButtons() {
-        for (_, button) in codeBlockCopyButtons {
-            button.removeFromSuperview()
-        }
-        codeBlockCopyButtons.removeAll()
-    }
-}
-
-// MARK: - Markdown Preview with WKWebView
-
-struct MarkdownPreviewView: NSViewRepresentable {
-    let markdownContent: String
-    let isDarkMode: Bool
-    let bodyFontFamily: String
-    let monoFontFamily: String
-    let fontSize: Int
-    let lineHeight: Double
-    var currentFileURL: URL? = nil
-    var onResolveWikilink: ((String) -> Void)? = nil
-    var onToggleCheckbox: ((Int) -> Void)? = nil
-
-    class Coordinator: NSObject, WKNavigationDelegate, WKScriptMessageHandler {
-        var parent: MarkdownPreviewView
-        var lastMarkdown: String?
-        var lastIsDarkMode: Bool?
-        var lastBodyFontFamily: String?
-        var lastMonoFontFamily: String?
-        var lastFontSize: Int?
-        var lastLineHeight: Double?
-        var lastFileURL: URL?
-        var pendingScrollY: CGFloat = 0
-
-        init(_ parent: MarkdownPreviewView) { self.parent = parent }
-
-        func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-            guard pendingScrollY > 0 else { return }
-            let y = pendingScrollY
-            pendingScrollY = 0
-            webView.evaluateJavaScript("window.scrollTo(0, \(y))") { _, _ in }
-        }
-
-        func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
-            guard let url = navigationAction.request.url else { decisionHandler(.allow); return }
-            if url.scheme == "wikilink" {
-                let destination = (url.host ?? url.path.trimmingCharacters(in: CharacterSet(charactersIn: "/"))).lowercased()
-                parent.onResolveWikilink?(destination)
-                decisionHandler(.cancel)
-                return
-            }
-            if url.scheme == "http" || url.scheme == "https" {
-                NSWorkspace.shared.open(url)
-                decisionHandler(.cancel)
-                return
-            }
-            // Allow file:// and about: (initial HTML load)
-            decisionHandler(.allow)
-        }
-
-        func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
-            if message.name == "toggleCheckbox", let offset = message.body as? Int {
-                parent.onToggleCheckbox?(offset)
-            }
-        }
-
-    }
-
-    func makeCoordinator() -> Coordinator { Coordinator(self) }
-
-    func makeNSView(context: Context) -> WKWebView {
-        let config = WKWebViewConfiguration()
-        config.preferences.javaScriptEnabled = true
-        config.userContentController.add(context.coordinator, name: "toggleCheckbox")
-        let webView = WKWebView(frame: .zero, configuration: config)
-        webView.setValue(false, forKey: "drawsBackground")
-        webView.navigationDelegate = context.coordinator
-        return webView
-    }
-
-    func updateNSView(_ webView: WKWebView, context: Context) {
-        guard markdownContent != context.coordinator.lastMarkdown ||
-              isDarkMode != context.coordinator.lastIsDarkMode ||
-              bodyFontFamily != context.coordinator.lastBodyFontFamily ||
-              monoFontFamily != context.coordinator.lastMonoFontFamily ||
-              fontSize != context.coordinator.lastFontSize ||
-              lineHeight != context.coordinator.lastLineHeight ||
-              currentFileURL != context.coordinator.lastFileURL else { return }
-        context.coordinator.parent = self
-        context.coordinator.lastMarkdown = markdownContent
-        context.coordinator.lastIsDarkMode = isDarkMode
-        context.coordinator.lastBodyFontFamily = bodyFontFamily
-        context.coordinator.lastMonoFontFamily = monoFontFamily
-        context.coordinator.lastFontSize = fontSize
-        context.coordinator.lastLineHeight = lineHeight
-        context.coordinator.lastFileURL = currentFileURL
-        let baseDir = currentFileURL?.deletingLastPathComponent()
-        let html = generateHTML(from: markdownContent, isDarkMode: isDarkMode, baseDir: baseDir)
-        // Save scroll position before reload, restore after load finishes
-        webView.evaluateJavaScript("window.scrollY") { scrollY, _ in
-            if let y = scrollY as? CGFloat, y > 0 {
-                context.coordinator.pendingScrollY = y
-            }
-        }
-        webView.loadHTMLString(html, baseURL: baseDir)
-    }
-
-    private func generateHTML(from markdown: String, isDarkMode: Bool, baseDir: URL? = nil) -> String {
-        var html = MarkdownPreviewRenderer().renderBody(from: markdown)
-        // Inline local images as data URIs so they render without file:// access
-        if let baseDir {
-            let imgRegex = try? NSRegularExpression(pattern: #"<img\s+src="([^"]+)""#)
-            let nsHTML = html as NSString
-            var replacements: [(NSRange, String)] = []
-            imgRegex?.enumerateMatches(in: html, range: NSRange(location: 0, length: nsHTML.length)) { match, _, _ in
-                guard let match, match.numberOfRanges > 1 else { return }
-                let srcRange = match.range(at: 1)
-                let src = nsHTML.substring(with: srcRange)
-                // Skip already-inlined or remote URLs
-                guard !src.hasPrefix("data:"), !src.hasPrefix("http://"), !src.hasPrefix("https://") else { return }
-                let imageURL = baseDir.appendingPathComponent(src)
-                guard let data = try? Data(contentsOf: imageURL) else { return }
-                let ext = imageURL.pathExtension.lowercased()
-                let mime: String
-                switch ext {
-                case "png": mime = "image/png"
-                case "jpg", "jpeg": mime = "image/jpeg"
-                case "gif": mime = "image/gif"
-                case "svg": mime = "image/svg+xml"
-                case "webp": mime = "image/webp"
-                default: mime = "application/octet-stream"
-                }
-                let dataURI = "data:\(mime);base64,\(data.base64EncodedString())"
-                replacements.append((srcRange, dataURI))
-            }
-            // Apply replacements in reverse order to preserve ranges
-            for (range, replacement) in replacements.reversed() {
-                html = (html as NSString).replacingCharacters(in: range, with: replacement)
-            }
-        }
-        
-        let textColor = isDarkMode ? "#E0E0E0" : "#333333"
-        let backgroundColor = isDarkMode ? "#1E1E1E" : "#FFFFFF"
-        let borderColor = isDarkMode ? "#444444" : "#CCCCCC"
-        let headerBgColor = isDarkMode ? "#2D2D2D" : "#F5F5F5"
-        let bodyFontStack = MarkdownPreviewCSS.bodyFontStack(for: bodyFontFamily)
-        let monoFontStack = MarkdownPreviewCSS.monoFontStack(for: monoFontFamily)
-        let bodyFontSize = MarkdownPreviewCSS.bodyFontSize(for: fontSize)
-        let tableFontSize = MarkdownPreviewCSS.tableFontSize(for: fontSize)
-        let codeFontSize = MarkdownPreviewCSS.codeFontSize(for: fontSize)
-        let bodyLineHeight = MarkdownPreviewCSS.lineHeight(for: lineHeight)
-        let h1Size = MarkdownPreviewCSS.headingFontSize(level: 1, baseSize: fontSize)
-        let h2Size = MarkdownPreviewCSS.headingFontSize(level: 2, baseSize: fontSize)
-        let h3Size = MarkdownPreviewCSS.headingFontSize(level: 3, baseSize: fontSize)
-        let h4Size = MarkdownPreviewCSS.headingFontSize(level: 4, baseSize: fontSize)
-        let h5Size = MarkdownPreviewCSS.headingFontSize(level: 5, baseSize: fontSize)
-        let h6Size = MarkdownPreviewCSS.headingFontSize(level: 6, baseSize: fontSize)
-        
-        return """
-        <!DOCTYPE html>
-        <html>
-        <head>
-            <meta charset="UTF-8">
-            <style>
-                * { margin: 0; padding: 0; box-sizing: border-box; }
-                body {
-                    font-family: \(bodyFontStack);
-                    font-size: \(bodyFontSize)px;
-                    line-height: \(bodyLineHeight);
-                    color: \(textColor);
-                    background-color: \(backgroundColor);
-                    margin: 0;
-                    padding: 20px;
-                }
-                table {
-                    border-collapse: collapse;
-                    width: 100%;
-                    margin: 16px 0;
-                    font-size: \(tableFontSize)px;
-                }
-                th, td {
-                    border: 1px solid \(borderColor);
-                    padding: 8px 12px;
-                    text-align: left;
-                }
-                th {
-                    background-color: \(headerBgColor);
-                    font-weight: 600;
-                }
-                tr:nth-child(even) {
-                    background-color: \(isDarkMode ? "#252525" : "#FAFAFA");
-                }
-                h1 { font-size: \(h1Size)px; margin: 24px 0 16px 0; font-weight: 600; }
-                h2 { font-size: \(h2Size)px; margin: 24px 0 16px 0; font-weight: 600; }
-                h3 { font-size: \(h3Size)px; margin: 20px 0 14px 0; font-weight: 600; }
-                h4 { font-size: \(h4Size)px; margin: 18px 0 12px 0; font-weight: 600; }
-                h5 { font-size: \(h5Size)px; margin: 16px 0 10px 0; font-weight: 600; }
-                h6 { font-size: \(h6Size)px; margin: 14px 0 8px 0; font-weight: 600; }
-                p { margin: 12px 0; }
-                p:empty { margin: 0; }
-                ul, ol {
-                    margin: 12px 0;
-                    padding-left: 1.5em;
-                }
-                ul ul, ul ol, ol ul, ol ol {
-                    margin: 2px 0;
-                }
-                li {
-                    margin: 4px 0;
-                }
-                code {
-                    background-color: \(isDarkMode ? "#2D2D2D" : "#F0F0F0");
-                    padding: 2px 6px;
-                    border-radius: 3px;
-                    font-family: \(monoFontStack);
-                    font-size: \(codeFontSize)px;
-                }
-                pre {
-                    background-color: \(isDarkMode ? "#2D2D2D" : "#F5F5F5");
-                    padding: 16px;
-                    border-radius: 6px;
-                    overflow-x: auto;
-                    margin: 16px 0;
-                }
-                pre code {
-                    background-color: transparent;
-                    padding: 0;
-                }
-                blockquote {
-                    border-left: 4px solid \(borderColor);
-                    margin: 12px 0;
-                    padding-left: 16px;
-                    color: \(isDarkMode ? "#AAAAAA" : "#666666");
-                }
-                .callout {
-                    border-left-color: \(isDarkMode ? "#6B9BFF" : "#0066CC");
-                    background: \(isDarkMode ? "rgba(107, 155, 255, 0.08)" : "rgba(0, 102, 204, 0.06)");
-                    border-radius: 8px;
-                    padding: 12px 14px;
-                    color: \(textColor);
-                }
-                .callout-title {
-                    font-weight: 700;
-                    margin-bottom: 6px;
-                }
-                .callout-body {
-                    color: \(textColor);
-                }
-                a {
-                    color: \(isDarkMode ? "#6B9BFF" : "#0066CC");
-                    text-decoration: none;
-                }
-                a:hover {
-                    text-decoration: underline;
-                }
-                a.wikilink {
-                    font-weight: 500;
-                }
-                .embed {
-                    display: inline-block;
-                    padding: 2px 8px;
-                    border-radius: 999px;
-                    background-color: \(isDarkMode ? "#2A2A2A" : "#EFEFEF");
-                    color: \(textColor);
-                }
-                .task-list {
-                    list-style: none;
-                    padding-left: 0;
-                }
-                .task-item {
-                    display: flex;
-                    align-items: baseline;
-                    gap: 8px;
-                }
-                .task-item input[type="checkbox"] {
-                    accent-color: \(isDarkMode ? "#6B9BFF" : "#0066CC");
-                    cursor: pointer;
-                    width: 14px;
-                    height: 14px;
-                    flex-shrink: 0;
-                    margin-top: 2px;
-                }
-                hr {
-                    border: none;
-                    border-top: 1px solid \(borderColor);
-                    margin: 24px 0;
-                }
-                img {
-                    max-width: 100%;
-                    height: auto;
-                    border-radius: 4px;
-                    display: block;
-                    margin: 8px 0;
-                }
-                strong { font-weight: 600; }
-                em { font-style: italic; }
-                del { text-decoration: line-through; }
-                \(SyntaxHighlightTheme.css(forDarkMode: isDarkMode))
-            </style>
-        </head>
-        <body>
-            \(html)
-        </body>
-        </html>
-        """
-    }
-}
